Changes in login page
